% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_env_array.R
\name{check_env_array}
\alias{check_env_array}
\title{Validate the environmental data array}
\usage{
check_env_array(env_dat, name = "env_dat")
}
\arguments{
\item{env_dat}{The object to validate.}

\item{name}{Character scalar.  Variable name shown in error messages.
Defaults to \code{"env_dat"}.}
}
\value{
\code{env_dat} invisibly (allows use in a pipe).
}
\description{
Checks that \code{env_dat} is a 3-dimensional array with dimensions
\code{n_loc x n_time x p} (number of locations x time-series length x number
of environmental variables) and contains no missing values.  Throws an
informative error if any condition is violated.
}
\keyword{internal}
