% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_equivalence_class.R
\name{convert_equivalence_class}
\alias{convert_equivalence_class}
\title{Converts a set of parameters to other representatives of the same equivalence
class}
\usage{
convert_equivalence_class(p, flip, perm)
}
\arguments{
\item{p}{Named list with entries mu, sigltil, sigrtil, ctil, pd, and o_mat}

\item{flip}{Vector of binaries corresponding to which columns of o_mat are to
have their sign change (which a concomitant switch of the corresponding
entries of sigltil and sigrtil). Length must equal the number of columns of
o_mat.}

\item{perm}{Permutation to be applied to the columns of o_mat.}
}
\value{
List with entries o_mat, sigltil, and sigrtil
}
\description{
Model parameters in the biological scale are only determined up to an
equivalence class. This function converts a set of parameters to another set
of equivalent parameters.
}
\examples{
convert_equivalence_class(
  p = math_to_bio(example_1$optim_par_vec),
  flip = c(1, 0),
  perm = c(1, 2)
)
}
