% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_1_bio01.R
\docType{data}
\name{example_1_bio01}
\alias{example_1_bio01}
\title{Time series of annual average temperature}
\format{
A raster time series with 39 years of bio01 data. Centered in
the south of New Mexico, United States.
}
\source{
<https://www.chelsa-climate.org/>
}
\usage{
example_1_bio01
}
\description{
A SpatRaster with with 128 x 123 cells and 39 layers representing
the annual average temperature for the years 1980-2018
}
\details{
The Annual average temperature is also known as the bioclimatic
variable 01 (bio1) Data was obtained from CHELSA 2.1. We compute
bio01 (Anual tas average) for a 30 years time series.
}
\keyword{datasets}
