% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_1_bio12.R
\docType{data}
\name{example_1_bio12}
\alias{example_1_bio12}
\title{Time series of annual precipitation}
\format{
A raster time series with 30 years of bio01 data
}
\source{
<https://www.chelsa-climate.org/>
}
\usage{
example_1_bio12
}
\description{
A SpatRaster with with 128 x 123 cells and 39 layers representing
the annual average temperature for the years 1970-1999.
}
\details{
The Annual precipitation is also known as the bioclimatic
variable 12 (bio12). Data was obtained from CHELSA v2.1. We compute bio12
(Anual precipitation) for a 3 years time series
}
\keyword{datasets}
