% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_1_env_array.R
\docType{data}
\name{example_1_env_array}
\alias{example_1_env_array}
\title{Environmental array}
\format{
An array with with dimensions 2000, 39 and 2.
}
\usage{
example_1_env_array
}
\description{
An array of a time series of 30 years of bio1 and bio12
bioclimatic variables
}
\details{
the dimension 2 correspond to 2000 sites, 39 time series steps
and 2 bioclimatic variables.
}
\keyword{datasets}
