% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_1_occurrence_df.R
\docType{data}
\name{example_1_occurrence_df}
\alias{example_1_occurrence_df}
\title{Occurence data frame of virtual species}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
  \item{\code{name}}{character Virtual species name: Mus virtualis}
  \item{\code{x}}{double latitude in the projection of the
  environmental layers of the example}
  \item{\code{y}}{double longitude in the projection of the
  environmental layers of the example}
  \item{\code{presence}}{double binary vector of present absence}
}
}
\usage{
example_1_occurrence_df
}
\description{
This is a data frame with latitude, longitude and presence or
absence binary value of a virtual species.
}
\details{
DETAILS
}
\keyword{datasets}
