% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_1_occurrence_vector.R
\docType{data}
\name{example_1_occurrence_vector}
\alias{example_1_occurrence_vector}
\title{Occurrence example}
\format{
A binary (ones and zeros) vector of occurrence of a virtual species.
}
\usage{
example_1_occurrence_vector
}
\description{
A vector of 2000 occurrences of a virtual species.
}
\details{
It helps to test the log likelihood function.
}
\keyword{datasets}
