% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{examples}
\alias{examples}
\title{Examples}
\format{
A list of 11 objects.
}
\usage{
examples
}
\description{
This is a list of examples objects for the xsdm package.
}
\details{
Includes 11 objects to be used innerly in the xsdm examples
}
\keyword{datasets}
