% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Functions to take the expit of numerical vectors.
expit exp(x)/(1 + exp(x))}
\usage{
expit(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
A real vector corresponding to the expits of x
}
\description{
Functions to take the expit of numerical vectors.
expit exp(x)/(1 + exp(x))
}
\examples{
expit(0)
expit(0.5)
expit(-1)
}
