% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit_general.R
\name{expit_general}
\alias{expit_general}
\title{Generalized expit function}
\usage{
expit_general(x, l_bound, u_bound, x0)
}
\arguments{
\item{x}{A real number to ma}

\item{l_bound}{Lower bound}

\item{u_bound}{Upper bound}

\item{x0}{Middle point}
}
\value{
A re-scaled valued between lower bound and upper bound with
the middle point x0
}
\description{
Generalized expit function
}
\examples{
expit_general(3, -pi, pi, 0)
}
