% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like_neg_ltsgr_r.R
\name{like_neg_ltsgr_r}
\alias{like_neg_ltsgr_r}
\title{Long-term stochastic growth rate worker function for the xsdm model, R version}
\usage{
like_neg_ltsgr_r(env_dat, mu, sigltil, sigrtil, o_mat)
}
\arguments{
\item{env_dat}{The environmental data array, dimensions
(number of locations) x (time series length) x (number of environmental
variables). Must not contain missing values.}

\item{mu}{Vector of optimal environmental values. Length \code{p=dim(env_dat)[3]}.
Unconstrained real numbers.}

\item{sigltil}{Vector specifying width of the growth-environment function.
Length \code{p=dim(env_dat)[3]}. Positive real numbers, \code{Inf} entries also allowed.}

\item{sigrtil}{Vector specifying width of the growth-environment function.
Length \code{p=dim(env_dat)[3]}. Positive real numbers, \code{Inf} entries also allowed.}

\item{o_mat}{An orthogonal matrix, dimensions \code{p} by \code{p}.}
}
\value{
A vector of length equal to the number of locations, as described
above.
}
\description{
Computes the negative of the long-term stochastic growth rate, plus
log(lambda_max), for the xsdm model, for each location. This is the R version
of a worker function, see also the accompanying C version, which should
produce identical results but faster.
}
\details{
Being an internal function, there is no error checking. Note that
env_dat must be a 3d array (not a matrix or a vector) even if one of its
dimensions is 1. And \code{o_mat} must be a matrix even when \code{p} is 1
(in that case it's a 1 x 1 matrix, but not a scalar).
}
\keyword{internal}
