% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit function. Inverse of the expit function.}
\usage{
logit(x)
}
\arguments{
\item{x}{A numeric value to scale from 0 to 1 to -inf to inf}
}
\value{
A numeric
}
\description{
Logit function. Inverse of the expit function.
}
\examples{
logit(0.5)
}
\keyword{internal}
