% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_bio.R
\name{loglik_bio}
\alias{loglik_bio}
\title{Log-likelihood function for the xsdm model, parameters on the biological
scale.}
\usage{
loglik_bio(
  env_dat,
  occ,
  mu,
  sigltil,
  sigrtil,
  o_mat,
  ctil,
  pd,
  return_prob = FALSE,
  sum_log_p = TRUE,
  num_threads = RcppParallel::defaultNumThreads()
)
}
\arguments{
\item{env_dat}{The environmental data array, dimensions
\code{n_loc x n_time x p} (number of locations x time-series length x number
of environmental variables). Must be a 3-dimensional array with no missing
values.}

\item{occ}{Presence/pseudo-absence binary vector. Same length as dimension 1 of
\code{env_dat}.}

\item{mu}{Vector of optimal environmental values. Length \code{p=dim(env_dat)[3]}.
Unconstrained real numbers.}

\item{sigltil}{Vector specifying width of the growth-environment function.
Length \code{p=dim(env_dat)[3]}. Positive real numbers, \code{Inf} entries also allowed.}

\item{sigrtil}{Vector specifying width of the growth-environment function.
Length \code{p=dim(env_dat)[3]}. Positive real numbers, \code{Inf} entries also allowed.}

\item{o_mat}{An orthogonal matrix, dimensions \code{p} by \code{p}.}

\item{ctil}{Scalar. Relates to the center of the detection-link function.}

\item{pd}{Maximum probability of detection of the species. Parameter between
0 and 1.}

\item{return_prob}{Logical (default FALSE). Flag to return likelihood instead
of log-likelihood.}

\item{sum_log_p}{Logical (default TRUE). If FALSE, returns the individual
log-likelihoods (or likelihoods, if \code{return_prob} is TRUE) associated with the
individual locations, instead of their sum (product, if \code{return_prob} is TRUE).}

\item{num_threads}{Number of threads for parallel computation. Defaults to
\code{RcppParallel::defaultNumThreads()}.}
}
\value{
A single value, the log-likelihood (or the likelihood, if \code{return_prob}
is TRUE); or a vector of location specific values of \code{sum_log_p} is FALSE.
}
\description{
Computes the log-likelihood for the xsdm model given environmental data, a
vector of occurrences and pseudo-absences, and model parameters on the
biological scale.
}
\examples{
ll <- loglik_bio(
  env_dat = examples$env_array,
  occ = examples$occ_vec,
  mu = examples$par_list$mu,
  sigltil = examples$par_list$sigltil,
  sigrtil = examples$par_list$sigrtil,
  o_mat = examples$par_list$o_mat,
  ctil = examples$par_list$ctil,
  pd = examples$par_list$pd
)
ll
}
