% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mask_names.R
\name{make_mask_names}
\alias{make_mask_names}
\title{Function to facilitate the creation of the argument \code{mask} to the function
\code{loglik_math}}
\usage{
make_mask_names(p)
}
\arguments{
\item{p}{A positive integer representing the number of environmental variables
to be used in the xsdm model, i.e., \code{dim(env_dat)[3]} for the
\code{env_dat} argument of the \code{loglik_math} function}
}
\value{
A named numeric vector full of NAs, with the names generated according
to the conventions in Details of the function \code{loglik_math}. See also Details
below.
}
\description{
The argument \code{mask} to the function \code{loglik_math} is required to
follow some very specific conventions in order to reduce the risk of errors
coming from mismatched arguments. This function facilitates the creation of
such vectors.
}
\details{
The output has length \code{3*p+(p^2-p)/2+2}. The names of the entries are
\code{mu1}, \code{mu2}, \ldots, \code{mup}, \code{sigltil1}, \code{sigltil2},
\ldots, \code{sigltilp}, \code{sigrtil1}, \code{sigrtil2}, \ldots,
\code{sigrtilp}, \code{o_pari} for \code{i} ranging from 1 to \code{(p^2-p)/2},
\code{ctil}, and \code{pd}. All entries are \code{NA}.
}
\examples{
make_mask_names(2)
}
