% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mll_df_example.R
\docType{data}
\name{mll_df_example}
\alias{mll_df_example}
\title{Fitted maximum liklihood example}
\format{
A data frame with fitted parameters, the
value of the likelihood function at numerical maximum
and the flag for convergency
}
\usage{
mll_df_example
}
\description{
A data.frame with 100 fitted optimizations
of the maximum likelihood
}
\details{
It helps to test the methods for fitted objects
}
\keyword{datasets}
