% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus_virtualis.R
\docType{data}
\name{mus_virtualis}
\alias{mus_virtualis}
\title{DATASET_TITLE}
\format{
A data frame with 20 rows and 4 variables:
\describe{
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{y}}{double COLUMN_DESCRIPTION}
  \item{\code{x}}{double COLUMN_DESCRIPTION}
  \item{\code{presence}}{double COLUMN_DESCRIPTION}
}
}
\usage{
mus_virtualis
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
