% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_list_example.R
\docType{data}
\name{param_list_example}
\alias{param_list_example}
\title{Parameter list example}
\format{
A list with parameters ctil , O, mu, sigLtil, sigRtil, pd
}
\usage{
param_list_example
}
\description{
A list with parameters to test the log likelihood function
}
\details{
This list is passed to functions to test it functionality
}
\keyword{datasets}
