% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_table_example.R
\docType{data}
\name{param_table_example}
\alias{param_table_example}
\title{Environmental parameter table}
\format{
A data.frame with with dimensions 9 columns and 100 rows:
The parameters are mu1, mu2, sigl1, sigl2, sigr1, sigr1, ctil, pd, and o_par1
}
\usage{
param_table_example
}
\description{
An data frame with a set of 100 parameters to test the
optimization of the likelihood function
}
\details{
This set of parameters correspond to the case of a two environmental
time series fit
}
\keyword{datasets}
