% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramListExample.R
\docType{data}
\name{paramListExample}
\alias{paramListExample}
\title{Parameter list example}
\format{
A list with parameters ctil , O, mu, sigLtil, sigRtil, pd
}
\usage{
paramListExample
}
\description{
A list with parameters to test the log likelihood function
}
\details{
This list is passed to functions to test it functionality
}
\keyword{datasets}
