% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_parms.R
\name{start_parms}
\alias{start_parms}
\title{Starting parameters for the optimization}
\usage{
start_parms(env_dat, mask = NULL, breadth = 1, num_starts = 100)
}
\arguments{
\item{env_dat}{The environmental array for only the observed occurrences}

\item{mask}{Either NULL or a named numeric vector. Names must be as specified
by calling \code{make_mask_names}. The NULL case means planned optimizations
will be over all model parameters, so start parameter sets should include all
model parameters. The non-NULL case means some parameters will not be
specified in the output of this function because the optimizations which
are planned will fix those parameters anyway. The most common case for most
applications will be \code{mask=NULL}.}

\item{breadth}{Scalar in \code{[0, 1]} controlling how wide the search
ranges are around their (fixed, data-driven) center. \code{breadth = 1}
(the default) reproduces the pre-v0.3 behaviour that corresponded to
\code{quant_vec = c(0.1, 0.5, 0.9)}; \code{breadth = 0} collapses
every range to essentially a single point, equivalent to
\code{quant_vec = c(0.5 - 1e-6, 0.5, 0.5 + 1e-6)}. Values in between
interpolate linearly.}

\item{num_starts}{The number of samples of the hypercube}
}
\value{
A data frame with samples for each parameter to optimize
}
\description{
Generate starting parameters for the optimization of the xsdm log-likelihood.
Starting points are constructed from the environmental conditions at observed
presences (where \code{occ == 1}) using a Latin hypercube design for the
parameters based on the Sobol' low-discrepancy sequence.
}
\details{
The bounds and center of the search range for a mu parameter are
based on the quantiles in quant_vec applied to all observations of that
environmental variable, over space and time. The relationship between
quant_vec and the width of the ranges selected for the other parameters
varies, but generally wider ranges in quant_vec produce wider ranges for
start parameters.
}
\examples{
env_dat <- example_1$env_array[example_1$occ_df$presence == 1, , ]
start_parms(env_dat)
start_parms(env_dat, mask = c(mu2 = 5, pd = 1))
}
