% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsdm-package.R
\docType{package}
\name{xsdm-package}
\alias{xsdm}
\alias{xsdm-package}
\title{xsdm: Demographic Approach to Species Distribution Model}
\description{
Integrates concepts of stochastic demography into species distribution modelling. The main approach maximizes a likelihood function based on environmental information and presence/absence records. This is used to reconstruct species' fundamental ecological niches and to project their potential geographic range. Data requirements include species presence/absence records and a timeseries of environmental data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://xsdm-project.github.io/xsdm/}
  \item \url{https://xsdm-project.github.io/xsdm-devel/}
  \item Report bugs at \url{https://github.com/xsdm-project/xsdm-devel/issues}
}

}
\author{
\strong{Maintainer}: Angel Luis Robles Fernandez \email{a.l.robles.fernandez@gmail.com} (\href{https://orcid.org/0000-0002-4674-4270}{ORCID})

Authors:
\itemize{
  \item Angel Luis Robles Fernandez \email{a.l.robles.fernandez@gmail.com} (\href{https://orcid.org/0000-0002-4674-4270}{ORCID})
  \item Emilio Berti \email{emilio.berti@idiv.de} (\href{https://orcid.org/0000-0001-9286-011X}{ORCID})
  \item Daniel C. Reuman \email{reuman@ku.edu} (\href{https://orcid.org/0000-0002-1407-8947}{ORCID})
}

}
