% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_between_params.R
\name{dist_between_params}
\alias{dist_between_params}
\title{Provides a measure of distance in parameter space between two sets of
parameters, using the Hungarian algorithm.}
\usage{
dist_between_params(p1, p2, mask = NULL, give_closest_rep = FALSE)
}
\arguments{
\item{p1}{First set of parameters. May be math- or biological-scale parameters,
following the associated formatting requirements in each case.}

\item{p2}{Second set of parameters, same format options as the first}

\item{mask}{Same format as the \code{mask} argument to \code{\link{loglik_math}}
and \code{\link{start_parms}}. Ignored if both \code{p1} and \code{p2} are on
the biological scale. Otherwise the parameters contained in \code{mask} must
exactly complement those contained in whichever of \code{p1} and \code{p2} are on the
math scale.}

\item{give_closest_rep}{TRUE if you also want the member of the equivalence class
of p1 which is closest to p2. Gives the answer on the biological scale.
Default FALSE.}
}
\value{
If \code{give_closest_rep} is FALSE then a single number which is the
distance. Otherwise a list with elements for the distance and for the
representative which gives that distance.
}
\description{
Computes a distance in parameter space between two sets of parameters. The
o_mat, sigltil, and sigrtil parameters are only determined up to equivalence
class. This function runs through all representatives of the equivalence
class of p1 and computes the distance to p2 for each, returning the minimum.
Optionally also returns the equivalence class representative that gives that
minimum. Only it does not literally compute all that, it uses the Hungarian
algorithm to speed it up. Distance is reckoned using sum-squared errors, on
the biological scale, except for sigltil and sigrtil which are inverted
before measuring distance because that is the scale it makes most sense to
measure distance for those parameters.
}
\examples{
example_1_optim_param_list <- math_to_bio(example_1_optim_param_vector)
example_1_optim_param_list_equivalent <- math_to_bio(example_1_optim_param_vector_equivalent)
dist_between_params(
  p1 = example_1_optim_param_list,
  p2 = example_1_optim_param_list_equivalent
)
dist_between_params(
  p1 = example_1_optim_param_vector,
  p2 = example_1_optim_param_vector_equivalent
)
}
