% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_data_array.R
\name{env_data_array}
\alias{env_data_array}
\title{Get an array of environmental data from presence-absence points.}
\usage{
env_data_array(env_data, occ = NULL)
}
\arguments{
\item{env_data}{List of environmental variables time series stacks (each a
SpatRaster with multiple layers).}

\item{occ}{Occurrence data frame. Should contain columns "name", "lon", "lat",
"presence". If NULL, returns data for all raster cells.}
}
\value{
A 3D array of dimensions M (points or cells) × N (time steps) × P
  (environmental variables). The first dimension has no dimnames; the second
  is named "time" with layer names from the first raster; the third is named
  "var" with the names of `env_data`.
}
\description{
Get an array of environmental data from presence-absence points.
}
\examples{
bio1_ts <- terra::unwrap(example_1$bio01)
bio12_ts <- terra::unwrap(example_1$bio12)
env_data <- list(bio1 = bio1_ts, bio12 = bio12_ts)
occ <- example_1$occ_df[1:5, ]
# Return array correspoding to each presence absence provided
env_data_array(env_data, occ)
# Return all the environmental in the rasters
env_data_array(env_data, occ)
}
