% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like_neg_ltsgr_cpp.R
\name{like_neg_ltsgr_cpp}
\alias{like_neg_ltsgr_cpp}
\title{Long-term stochastic growth rate worker function for the xsdm model, version
using C++ code}
\usage{
like_neg_ltsgr_cpp(
  env_dat,
  mu,
  sigltil,
  sigrtil,
  o_mat,
  num_threads = RcppParallel::defaultNumThreads()
)
}
\arguments{
\item{env_dat}{The environmental data array, dimensions
(number of locations) x (time series length) x (number of environmental
variables). Must not contain missing values.}

\item{mu}{Vector of optimal environmental values.
Length \code{p=dim(env_dat)[3]}. Unconstrained real numbers.}

\item{sigltil}{Vector specifying width of the growth-environment function.
Length \code{p=dim(env_dat)[3]}. Positive real numbers, \code{Inf} entries
also allowed.}

\item{sigrtil}{Vector specifying width of the growth-environment function.
Length \code{p=dim(env_dat)[3]}. Positive real numbers, \code{Inf} entries
also allowed.}

\item{o_mat}{An orthogonal matrix, dimensions \code{p} by \code{p}.}

\item{num_threads}{Number of threads for parallel computation. Defaults to
\code{RcppParallel::defaultNumThreads()}.}
}
\value{
A vector of length equal to the number of locations, as described
above.
}
\description{
Computes the negative of the long-term stochastic growth rate, plus
log(lambda_max), for the xsdm model, for each location. This is the fast
version  of a worker function that relies on C++ code. See also the
accompanying R version, which should produce identical results, but slower,
and is included for testing. This c++ version is optimized
using \pkg{RcppParallel}.
}
\details{
Internally, this function:
\enumerate{
  \item Reshapes the environmental data into a matrix.
  \item Computes inverse (diagonal) matrices for asymmetry adjustments.
  \item Calls the C++ function \code{like_ltsg()} for efficient likelihood
  computation.
}
}
\note{
Ensure that `env_dat` has no missing values. The parameter vectors
`mu`, `sigltil`, and `sigrtil` must have length equal to the
number of environmental variables (`p`). The argument `o_mat` must
be a `p` x `p` orthogonal matrix (i.e., `o_mat %*% o_mat` is
the identity).
}
\examples{
# Example usage:
like_neg_ltsgr_cpp(env_dat = examples$env_array,
                  mu = examples$par_list$mu,
                  sigltil = examples$par_list$sigltil,
                  sigrtil = examples$par_list$sigrtil,
                  o_mat = examples$par_list$o_mat)
}
