% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp.R
\name{vsp}
\alias{vsp}
\title{Generate a virtual species probability map}
\usage{
vsp(env_data, param_list, return_raster)
}
\arguments{
\item{env_data}{A named list of time-series raster objects (e.g., bioclimatic
variables). Each element should be a `SpatRaster` or similar object from the
`terra` package.}

\item{param_list}{A named list of parameters required by `log_prob_detect()`.
Must include `mu`, `sigltil`, `sigrtil`, `ctil`, `pd`, and `o_mat`.}

\item{return_raster}{Logical. If `TRUE`, returns a `SpatRaster` object with
probabilities. If `FALSE`, returns a tibble with columns `x`, `y`, and
`probs`.}
}
\value{
Either:
  * A `SpatRaster` object (if `return_raster = TRUE`), or
  * A tibble with coordinates and probability values
  (if `return_raster = FALSE`).
}
\description{
Creates a virtual species probability-of-detection map based on environmental
time-series data and a set of species-specific parameters.
}
\details{
Internally, the function:
\enumerate{
  \item Converts the list of rasters into an array using `env_data_array()`.
  \item Applies `log_prob_detect()` with the provided parameters.
  \item Exponentiates the log-probabilities to obtain detection
  probabilities.
}
}
\examples{
# Example using chelsa dataset preloaded in xsdm package:
bio1_ts <- terra::unwrap(example_1_bio01)
bio12_ts <- terra::unwrap(example_1_bio12)
bio1_ts <- bio1_ts / 100
bio12_ts <- bio12_ts / 100
env_data <- list(bio1 = bio1_ts, bio12 = bio12_ts)
param_list_example <- list(
  mu = c(14, 6.5),
  sigltil = c(0.46, 1.08),
  sigrtil = c(0.105, 0.9),
  ctil = -18.14,
  pd = 0.89,
  o_mat = matrix(c(-0.18, 0.983, -0.983, -0.18), nrow = 2, ncol = 2)
)
vsp(env_data, param_list_example, return_raster = TRUE)

}
\seealso{
[env_data_array()], [log_prob_detect()], [terra::rast()]
}
